[I-126] Do not set default country code
"US" was being used as a default country code if one was not
present in the pod annotations. Remove these defauts and allow
CC not present to be valid.

Also force input CC values to lower case as the salad-client tests
only consider lower case values.  Note that some of the default values
being removed here were in upper case, so ignoted anyway.

Signed-off-by: Dean Troyer <dean.troyer@salad.com>